+++ .gitignore
@@ -41,6 +41,7 @@ yarn-error.log*
 # typescript
 *.tsbuildinfo
 next-env.d.ts
+next.config.mjs
 
 # uploads
 /src/temp/ocr/*
