+++ src/lib/ocr/server/services/index.js
@@ -1,0 +1,6 @@
+// src/lib/ocr/server/services/index.js
+export { uploadService } from './uploadService';
+export { storageService } from './storageService';
+export { qrService } from './qrService';
+export { transmittalService } from './transmittalService';
+export { cleanupService } from './cleanupService';
